Renamed files and components
